refactor: change the basic structure and remove dotenv support
