+++ package.json
@@ -3,16 +3,16 @@
   "version": "1.0.0",
   "private": "true",
   "scripts": {
-    "start": "node app.js",
-    "dev": "nodemon app.js",
-    "lint": "eslint"
+    "start": "node ./src/server.js",
+    "dev": "nodemon ./src/server.js",
+    "lint": "eslint",
+    "postinstall": "prisma generate"
   },
   "prisma": {
     "seed": "node prisma/seeders/seed.js"
   },
   "dependencies": {
     "@prisma/client": "^5.15.0",
-    "dotenv": "^16.4.5",
     "express": "^4.19.2"
   },
   "devDependencies": {
