+++ src/app.js
@@ -1,0 +1,12 @@
+const express = require("express");
+
+const app = express();
+
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
+app.get("/", async (req, res) => {
+  res.send("♻️ Recyclelify");
+});
+
+module.exports = app;
