+++ src/server.js
@@ -1,0 +1,21 @@
+const app = require("./app");
+
+const port = process.env.PORT || 3000;
+
+app.listen(port, async () => {
+  console.log("🚀 Server running on port", port);
+});
+
+process.on("unhandledRejection", async (reason, promise) => {
+  console.group("Unhandled Rejection");
+  console.error("reason:", reason);
+  console.error("at:", promise);
+  console.groupEnd();
+});
+
+process.on("uncaughtException", async (error, origin) => {
+  console.group("Uncaught Exception");
+  console.error("origin:", origin);
+  console.error("error:", error);
+  console.groupEnd();
+});
